Add test for PengRobinson calculation for mixture

+++ include/PhaseBehavior/Utilities/Types.hpp
@@ -11,7 +11,7 @@ namespace PhaseBehavior::Types {
      * This type represents the level of numerical precision used
      * throughout the codebase.
      */
-    using NumericalPrecision = double;
+    using NumericalPrecision = long double;
     using ConstexprConstant = decltype(Math::constants::pi<NumericalPrecision>);   
 }
